Fixed bug where the reset camera settings would reset to hardcoded defaults. Then would display defaults as floats. Now displays as integers.

+++ camera_setup_window.py
@@ -40,8 +40,8 @@ class CameraSetupWindow :
 
         # Camera settings
         self.current_profile = tk.StringVar(value="normal")
-        self.exposure_value = tk.DoubleVar(value=0)
-        self.gain_value = tk.DoubleVar(value=0)
+        self.exposure_value = tk.IntVar(value=0)
+        self.gain_value = tk.IntVar(value=0)
         # Sliders are currently being loaded in from file
         self.loading_sliders = False
         self.settings_status = tk.StringVar(value="Settings: Loaded normal profile")
@@ -320,8 +320,8 @@ class CameraSetupWindow :
         """
 
         return {
-            "exposure": self.exposure_value.get(),
-            "gain": self.gain_value.get(),
+            "exposure": int(self.exposure_value.get()),
+            "gain": int(self.gain_value.get()),
         }
     
     def apply_slider_settings_to_camera(self) :
